rename types/logger-lever.enum.ts -> types/logger-level.enum.ts create ng-package.json & angular.json file for ng-packagr build (ng6) remove gulpfile & gulp related dependency remove build:watch npm command (ng6 not support yet)
