+++ package.json
@@ -2,8 +2,7 @@
   "name": "ngx-logger",
   "version": "2.2.4",
   "scripts": {
-    "build": "gulp build",
-    "build:watch": "gulp",
+    "build": "ng build --project ngx-logger",
     "docs": "npm run docs:build",
     "docs:build": "compodoc -p tsconfig.json -n ngx-logger -d docs --hideGenerator",
     "docs:serve": "npm run docs:build -- -s",
@@ -36,44 +35,41 @@
     "url": "https://github.com/dbfannin/ngx-logger/issues"
   },
   "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@compodoc/compodoc": "^1.0.0-beta.10",
+    "@angular-devkit/build-ng-packagr": "^0.5.6",
+    "@angular/cli": "^6.0.0-rc.4",
+    "@angular/common": "^6.0.0-rc.4",
+    "@angular/compiler": "^6.0.0-rc.4",
+    "@angular/compiler-cli": "^6.0.0-rc.4",
+    "@angular/core": "^6.0.0-rc.4",
+    "@angular/platform-browser": "^6.0.0-rc.4",
+    "@angular/platform-browser-dynamic": "^6.0.0-rc.4",
+    "@compodoc/compodoc": "^1.1.2",
     "@types/jasmine": "2.5.53",
-    "@types/node": "~6.0.60",
+    "@types/node": "^6.0.106",
     "codelyzer": "~3.2.0",
     "concurrently": "^3.4.0",
-    "core-js": "^2.4.1",
-    "fs-extra": "^5.0.0",
-    "gulp": "^3.9.1",
-    "gulp-rename": "^1.2.2",
-    "gulp-rollup": "^2.15.0",
+    "core-js": "^2.5.5",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
     "karma-chrome-launcher": "~2.1.1",
     "karma-cli": "~1.0.1",
     "karma-coverage-istanbul-reporter": "^1.2.1",
-    "karma-jasmine": "~1.1.0",
+    "karma-jasmine": "^1.1.1",
     "karma-jasmine-html-reporter": "^0.2.2",
     "karma-typescript": "^3.0.12",
     "karma-typescript-es6-transform": "^1.0.4",
+    "ng-packagr": "^2.4.2",
     "node-sass": "^4.5.2",
     "node-sass-tilde-importer": "^1.0.0",
-    "node-watch": "^0.5.2",
+    "node-watch": "^0.5.8",
     "protractor": "~5.1.2",
-    "rollup": "^0.49.3",
-    "run-sequence": "^1.2.2",
-    "rxjs": "^5.5.2",
+    "rxjs": "^6.0.0-uncanny-rc.7",
     "systemjs": "^0.20.12",
     "ts-node": "~3.2.0",
     "tslint": "~5.7.0",
-    "typescript": "~2.4.2",
-    "zone.js": "^0.8.14"
+    "typescript": "~2.7.2",
+    "zone.js": "^0.8.26"
   },
   "engines": {
     "node": ">=6.0.0"
