+++ src/http-meta-data.interface.ts
@@ -1,4 +1,4 @@
-import {NgxLoggerLevel} from './types/logger-lever.enum';
+import {NgxLoggerLevel} from './types/logger-level.enum';
 
 export class HttpMetaDataInterface {
   level: NgxLoggerLevel;
