+++ src/index.ts
@@ -3,22 +3,23 @@ import {HttpClientModule} from '@angular/common/http';
 import {ModuleWithProviders, NgModule} from '@angular/core';
 
 import {NGXLogger} from './logger.service';
+import {LoggerConfig} from './logger.config';
+import {CustomNGXLoggerService} from './custom-logger.service';
+import {NGXLoggerHttpService} from './http.service';
+
 export * from './logger.service.mock';
 export * from './logger.service';
 
-import {LoggerConfig} from './logger.config';
 export * from './logger.config';
 
-import {CustomNGXLoggerService} from './custom-logger.service';
 export * from './custom-logger.service';
 export * from './custom-logger.service.mock';
 
-import {NGXLoggerHttpService} from './http.service';
 export * from './http.service';
 export * from './http.service.mock';
 
 export * from './utils/logger.utils';
-export * from './types/logger-lever.enum';
+export * from './types/logger-level.enum';
 export * from './http-meta-data.interface';
 
 @NgModule({
