+++ src/ng-package.json
@@ -1,0 +1,8 @@
+{
+  "$schema": "../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../dist",
+  "deleteDestPath": true,
+  "lib": {
+    "entryFile": "index.ts"
+  }
+}
